fix(striker-ui): correct show controls condition in server previews
