+++ striker-ui/components/Dashboard/ServerPanels.tsx
@@ -1,71 +1,8 @@
 import Grid from '@mui/material/Grid';
-import styled from '@mui/material/styles/styled';
 import { useMemo } from 'react';
 
-import {
-  Preview,
-  PreviewBox as BasePreviewBox,
-  PreviewFrame,
-} from '../Display';
 import Divider from '../Divider';
-import Link from '../Link';
-import { BodyText } from '../Text';
-
-const PreviewBox = styled(BasePreviewBox)(({ theme }) => {
-  const getHeight = (width = '0') => `calc(${width} * 0.6)`;
-
-  return {
-    width: '100%',
-
-    [theme.breakpoints.up('xs')]: {
-      height: getHeight('100vw'),
-    },
-
-    [theme.breakpoints.up('sm')]: {
-      height: getHeight('50vw'),
-    },
-
-    [theme.breakpoints.up('md')]: {
-      height: getHeight('100vw / 3'),
-    },
-
-    [theme.breakpoints.up('lg')]: {
-      height: getHeight('25vw'),
-    },
-
-    [theme.breakpoints.up('xl')]: {
-      height: getHeight('100vw / 6'),
-    },
-  };
-});
-
-const buildPreview = (server: APIServerOverview): React.ReactNode => (
-  <Grid item>
-    <PreviewFrame<APIServerOverview>
-      getHeader={({ jobs, name }) => {
-        if (jobs) {
-          return <BodyText noWrap>{name}</BodyText>;
-        }
-
-        return (
-          <Link href={`/server?name=${name}`} noWrap>
-            {name}
-          </Link>
-        );
-      }}
-      key={`${server.uuid}-preview`}
-      server={server}
-      showControls={!server.jobs}
-    >
-      <Preview
-        server={server}
-        slots={{
-          screenshotBox: <PreviewBox />,
-        }}
-      />
-    </PreviewFrame>
-  </Grid>
-);
+import ServerPreview from './ServerPreview';
 
 const ServerPanels: React.FC<ServerPanelsProps> = (props) => {
   const { groups, servers } = props;
@@ -84,7 +21,7 @@ const ServerPanels: React.FC<ServerPanelsProps> = (props) => {
 
           elements.push(
             <Grid key={`${uuid}-panel`} item xs={1}>
-              {buildPreview(server)}
+              <ServerPreview server={server} />
             </Grid>,
           );
         });
